Commit final

+++ src/modules/products/infra/typeorm/entities/Product.ts
@@ -20,7 +20,7 @@ class Product {
   @Column('decimal', { precision: 10, scale: 2 })
   price: number;
 
-  @Column('decimal')
+  @Column('int')
   quantity: number;
 
   @OneToMany(() => OrdersProducts, ordersProducts => ordersProducts.product)
